+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/RobotMetadataStatementGlobalSetting.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface RobotMetadataStatementGlobalSetting extends RobotGlobalSettingStatement {
+public interface RobotMetadataStatementGlobalSetting extends RobotGlobalSettingStatement, GlobalSettingStatementExpression {
 
   @NotNull
   List<RobotPositionalArgument> getPositionalArgumentList();