+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/RobotTemplateStatementsGlobalSetting.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface RobotTemplateStatementsGlobalSetting extends RobotGlobalSettingStatement {
+public interface RobotTemplateStatementsGlobalSetting extends RobotGlobalSettingStatement, GlobalSettingStatementExpression {
 
   @NotNull
   RobotKeywordCall getKeywordCall();